purge input
