+++ Game.lua
@@ -28,7 +28,7 @@ end
 local Game = class(
   function(self)
     self.scores = require("scores")
-    self.input = require("input")
+    self.input = { maxConfigurations = 8, inputConfigurations = {}}
     self.match = nil -- Match - the current match going on or nil if inbetween games
     self.battleRoom = nil -- BattleRoom - the current room being used for battles
     self.focused = true -- if the window is focused
